refactor: after review
