+++ frontend/src/components/workout/common/StickyHeader/types.ts
@@ -1,0 +1,10 @@
+import type { ChangeEvent, ReactNode } from 'react';
+
+export interface IStickyHeaderProps {
+  name: string;
+  date: string;
+  handleChange?: (e: ChangeEvent<HTMLInputElement>) => void;
+  readOnly?: boolean;
+  nameError?: string;
+  actions?: ReactNode;
+}
